backend/pyproject.toml
[tool.poetry]
name = "sonneshop"
version = "0.1.0"
description = ""
authors = ["Viktor Bagrov <Keramogranitus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
fastapi = "^0.115.0"
pydantic-settings = "^2.5.2"
pydantic = {extras = ["email"], version = "^2.9.2"}
pyjwt = "^2.9.0"
passlib = "^1.7.4"
redis = "^5.0.8"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
alembic = "^1.13.3"


[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.20.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
faker = "^30.0.0"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
